social-network-hw3/api-gateway-posts/requirements.txt
fastapi
uvicorn
pydantic
grpcio
protobuf
grpcio-tools
python-jose